exports page types const into own file
