+++ src/app/routes/utils/constructPageFetchUrl/page-types.ts
@@ -1,0 +1,9 @@
+const PAGE_TYPES = {
+  ARTICLE: 'article',
+  CPS_ASSET: 'cpsAsset',
+  HOME: 'home',
+  TOPIC: 'topic',
+  LIVE: 'live',
+} as const;
+
+export default PAGE_TYPES;
